Added config parameter for ignoring paths
Inside the config.xml there will be entries for ignored paths.
Every synchronized repository could contain a new xml element
describing an ignored folder like this:
...
<folder>
	...
	<ignoredFolder path="path/to/temp" />
	<ignoredFolder path="path/to/trash" />
	...
</folder>
...

This xml element is optional, so compatibility with old configs
is ensured.

The check, if a folder is ignored is not active at the moment. It
has to be evaluated, before enabling the checks for productivity.