Added alternative log4net config file path
If a log4net.config file exists in the path of the program config.xml, this file is
chosen as config file for log4net. The file also is watched by log4net which
can be used to change logging behaviours while a process is running. This could
be very helpful for support and development reasons.